Add a composer.json file, make autoloading work with Composer, add a test bootstrap and phpunit.xml.dist file for more convenience when running tests.
