+++ .gitignore
@@ -1,3 +1,5 @@
 pkg/
 coverage
 .DS_Store
+vendor/
+composer.lock
