+++ phpunit.xml.dist
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<phpunit bootstrap="test/bootstrap.php" colors="true">
+    <testsuites>
+        <testsuite name="Urlcrypt Test Suite">
+            <directory>test/Urlcrypt/</directory>
+        </testsuite>
+    </testsuites>
+
+    <filter>
+        <whitelist>
+            <directory suffix=".php">lib/Urlcrypt/</directory>
+        </whitelist>
+    </filter>
+</phpunit>
