+++ test/bootstrap.php
@@ -1,0 +1,4 @@
+<?php
+
+$loader = require __DIR__ . "/../vendor/autoload.php";
+$loader->addPsr4('Urlcrypt\\', __DIR__.'/Urlcrypt/');
